+++ src/main/java/net/geforcemods/securitycraft/datagen/BlockTagGenerator.java
@@ -578,10 +578,8 @@ public class BlockTagGenerator extends BlockTagsProvider {
 			}
 
 			switch (block) {
-				case ReinforcedSlabBlock slab when !woodenSlabs.contains(block) ->
-						tag(SCTags.Blocks.REINFORCED_SLABS).add(block);
-				case ReinforcedStairsBlock stairs when !woodenStairs.contains(block) ->
-						tag(SCTags.Blocks.REINFORCED_STAIRS).add(block);
+				case ReinforcedSlabBlock slab when !woodenSlabs.contains(block) -> tag(SCTags.Blocks.REINFORCED_SLABS).add(block);
+				case ReinforcedStairsBlock stairs when !woodenStairs.contains(block) -> tag(SCTags.Blocks.REINFORCED_STAIRS).add(block);
 				case ReinforcedWallBlock wall -> tag(SCTags.Blocks.REINFORCED_WALLS).add(block);
 				default -> {
 				}
@@ -973,7 +971,8 @@ public class BlockTagGenerator extends BlockTagsProvider {
 				SCContent.REINFORCED_RESIN_BRICK_STAIRS.get(),
 				SCContent.REINFORCED_CHISELED_RESIN_BRICKS.get(),
 				SCContent.REINFORCED_STONE_BUTTON.get(),
-				SCContent.REINFORCED_POLISHED_BLACKSTONE_BUTTON.get())
+				SCContent.REINFORCED_POLISHED_BLACKSTONE_BUTTON.get(),
+				SCContent.REINFORCED_LIGHTNING_ROD.get())
 		.addTag(SCTags.Blocks.REINFORCED_WALLS)
 		.addTag(SCTags.Blocks.REINFORCED_CAULDRONS);
 		tag(BlockTags.MINEABLE_WITH_SHOVEL).add(
@@ -1062,7 +1061,8 @@ public class BlockTagGenerator extends BlockTagsProvider {
 				SCContent.REINFORCED_COPPER_BULB.get(),
 				SCContent.REINFORCED_EXPOSED_COPPER_BULB.get(),
 				SCContent.REINFORCED_WEATHERED_COPPER_BULB.get(),
-				SCContent.REINFORCED_OXIDIZED_COPPER_BULB.get());
+				SCContent.REINFORCED_OXIDIZED_COPPER_BULB.get(),
+				SCContent.REINFORCED_LIGHTNING_ROD.get());
 	}
 
 	@Override