+++ Changelog.md
@@ -1,3 +1,8 @@
+--------------------------Changelog for the next version of SecurityCraft--------------------------
+
+- New: Reinforced Lightning Rod
+- Misc.: The minimum required NeoForge version is now 21.8.10
+
 --------------------------Changelog for v1.10-beta1 of SecurityCraft--------------------------
 
 - Fix: Crash when pushing blocks with reinforced pistons